chore(v2): clean some old files that are breaking tests
We will reimplement all of them ultimately, no big deal. If your PR is
about one of those widgets that was removed, just take your own files
and don't bother merging anything.

I cleaned as much as I could, I had to remove some tests of widgets that
are no more working because based on the old store etc..

Right now the widgets that are really finished are:
- Hits
- HitsPerPage
- Stats