Stop false positives!
This comp misfires a lot when loading legit results, just watch searchStore.results.length to find out.  This tweak at least makes it so it will only show when you actually launched a query.  There may be other things that would work (query maybe?) but this one seems to do the job for me.