fix(SearchBox): propagate the classname overrides to children
We never noticed it because this wasn't a very popular API, but if you use vanilla ais-search-box, you can not override the class names of the input or the clear which are nested within the component.

This is not an issue we have in v2, since we removed this API and no longer nest components in eachother, so it doesn't need to be passed around.

Using the `:class-names` API to pretend v1 is using InstantSearch.css with @francoischalifour made me realise that this is a useful API to keep, so it should be replicated in v2 as well.