feat(StateResults): give access to state
This changes the internal `connectStateResults` connector so that it exposes the state too, but without the methods.

This is not a breaking change in the widget, so should be straightforward to migrate to the new usage:

```diff
<ais-state-results>
- <template slot-scope="{ query, hits }">
+ <template slot-scope="{ results: { query, hits } }">
  </template>
</ais-state-results>
```

It allows you to read from the state as well, avoiding race conditions if for example hiding hits based on the query.

```diff
<ais-state-results>
- <div slot-scope="{ query }">
+ <div slot-scope="{ state: { query } }">
    <ais-hits v-if="query" />
  </div>
</ais-state-results>
```