fix(ssr): clearer error message when using ssr without the component
fixes #697

We are patching the InstantSearch instance in SSR, but you could be using the code for SSR without using the patched InstantSearch instance.

The error being thrown then is confusing, since it just complains about a missing property.

We have no differential error management with short & long error messages in prod/dev here yet, so I tried to be between both (not too long)