feat(core): expose source files
This should allow people who are using Webpack, which doesn't correctly work with "non-fully-used" single files to still import only what's needed.

The size of tree-shaking vue-instantsearch isn't too different from the whole bundle, because the biggest parts are the InstantSearch class & the algoliasearch-helper. Luckily this will be better in the next major version.

see #725 (fixes part of it as it gives a workaround, but ideally it wouldn't be necessary)